+++ CHANGES.md
@@ -1,7 +1,7 @@
 # GrooveJames change history
 
 ### r31 (unreleased)
--
+- bugfix: workaround for a problem in the blinkenlights MP3 IDv1 tagger: songs with unicode codepoints > 256 in the first 30 characters of its name could not be properly tagged, leading to NegativeArraySizeException and leftover files named "id3.<hash>.tmp" in the download folder
 
 ### r31 (2016/04/05)
 - bugfix: deal properly with songs without download url